+++ src/com/kaylerrenslow/armaDialogCreator/gui/fx/main/ADCMenuBar.java
@@ -23,10 +23,7 @@ import com.kaylerrenslow.armaDialogCreator.gui.fx.main.actions.mainMenu.edit.Edi
 import com.kaylerrenslow.armaDialogCreator.gui.fx.main.actions.mainMenu.edit.EditRedoAction;
 import com.kaylerrenslow.armaDialogCreator.gui.fx.main.actions.mainMenu.edit.EditUndoAction;
 import com.kaylerrenslow.armaDialogCreator.gui.fx.main.actions.mainMenu.edit.EditViewChangesAction;
-import com.kaylerrenslow.armaDialogCreator.gui.fx.main.actions.mainMenu.file.FileCloseProjectAction;
-import com.kaylerrenslow.armaDialogCreator.gui.fx.main.actions.mainMenu.file.FileExportAction;
-import com.kaylerrenslow.armaDialogCreator.gui.fx.main.actions.mainMenu.file.FileOpenAction;
-import com.kaylerrenslow.armaDialogCreator.gui.fx.main.actions.mainMenu.file.FileSaveAction;
+import com.kaylerrenslow.armaDialogCreator.gui.fx.main.actions.mainMenu.file.*;
 import com.kaylerrenslow.armaDialogCreator.gui.fx.main.actions.mainMenu.help.WikiUrlAction;
 import com.kaylerrenslow.armaDialogCreator.gui.fx.main.actions.mainMenu.view.*;
 import com.kaylerrenslow.armaDialogCreator.gui.img.ImagePaths;
@@ -47,7 +44,8 @@ class ADCMenuBar extends MenuBar {
 	private final MenuItem file_save = addOnAction(new MenuItem(MainMenuBar.FILE_SAVE), new FileSaveAction());
 	private final MenuItem file_export = addOnAction(new MenuItem(MainMenuBar.FILE_EXPORT), new FileExportAction());
 	private final MenuItem file_closeProject = addOnAction(new MenuItem(MainMenuBar.FILE_CLOSE_PROJECT), new FileCloseProjectAction());
-	
+	private final MenuItem file_exit = addOnAction(new MenuItem(MainMenuBar.FILE_EXIT), new FileExitAction());
+
 	
 	/*Edit*/
 	private final MenuItem edit_undo = new MenuItem(MainMenuBar.EDIT_UNDO, new ImageView(ImagePaths.ICON_UNDO));
@@ -115,7 +113,7 @@ class ADCMenuBar extends MenuBar {
 	/*help*/
 	private final MenuItem help_wiki = addOnAction(new MenuItem(MainMenuBar.HELP_WIKI), new WikiUrlAction());
 
-	private final Menu menuFile = new Menu(MainMenuBar.FILE, null, file_open, file_save, file_export, new SeparatorMenuItem(), file_closeProject);
+	private final Menu menuFile = new Menu(MainMenuBar.FILE, null, file_open, file_save, file_export, new SeparatorMenuItem(), file_closeProject, file_exit);
 	private final Menu menuEdit = new Menu(MainMenuBar.EDIT, null, edit_undo, edit_redo, edit_viewChanges, new SeparatorMenuItem(), edit_macros, edit_exportConfig);
 	private final Menu menuView = new Menu(MainMenuBar.VIEW, null, view_preview, view_showGrid, view_colors, view_darkTheme, backgroundAll, view_absRegionAll, view_ui, view_fullScreen);
 	private final Menu menuSettings = new Menu(MainMenuBar.SETTINGS, null, settings_configureDirs);
@@ -124,7 +122,7 @@ class ADCMenuBar extends MenuBar {
 
 	ADCMenuBar() {
 		this.getMenus().addAll(menuFile, menuEdit, menuView, menuSettings, menuCreate, menuHelp);
-		if(ArmaDialogCreator.getLaunchParameters().getUnnamed().contains(ArmaDialogCreator.ProgramArguments.ShowDebugFeatures.argText)){
+		if(ArmaDialogCreator.containsUnamedLaunchParameter(ArmaDialogCreator.ProgramArgument.ShowDebugFeatures)){
 			this.getMenus().add(new Menu("Test", null, addOnAction(new MenuItem("Test"), new TestAction())));
 		}
 	}